# specs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/products": {
      "get": {
        "description": "Browse the public product catalogue with optional filters for collection, category, price range, and sort order.",
        "operationId": "ProductsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "collectionSlug",
            "required": false,
            "in": "query",
            "schema": {
              "example": "summer-2025",
              "type": "string"
            }
          },
          {
            "name": "collectionId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "categorySlug",
            "required": false,
            "in": "query",
            "schema": {
              "example": "clothing",
              "type": "string"
            }
          },
          {
            "name": "categoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "subCategorySlug",
            "required": false,
            "in": "query",
            "schema": {
              "example": "t-shirts",
              "type": "string"
            }
          },
          {
            "name": "subCategoryId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "blue widget",
              "type": "string"
            }
          },
          {
            "name": "minPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "example": 10,
              "type": "number"
            }
          },
          {
            "name": "maxPrice",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 0,
              "example": 200,
              "type": "number"
            }
          },
          {
            "name": "sortBy",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "createdAt",
                "basePrice",
                "name"
              ]
            }
          },
          {
            "name": "sortOrder",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "summary": "List published products",
        "tags": [
          "Products"
        ]
      }
    },
    "/products/{name}": {
      "get": {
        "description": "Fetch a single published product by its URL slug, including all variants, photos, and category info.",
        "operationId": "ProductsController_findOne",
        "parameters": [
          {
            "name": "name",
            "required": true,
            "in": "path",
            "description": "Product URL slug (e.g. oversized_hoodie)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProductResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Product not found"
          }
        },
        "summary": "Get a product by slug",
        "tags": [
          "Products"
        ]
      }
    },
...
    "/admin/products/{id}": {
      "get": {
        "operationId": "AdminProductsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Get a product by ID (admin view with all fields)",
        "tags": [
          "Admin – Products"
        ]
      },
      "patch": {
        "operationId": "AdminProductsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProductDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminProductResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Update a product",
        "tags": [
          "Admin – Products"
        ]
      },
      "delete": {
        "operationId": "AdminProductsController_softDelete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Product UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Product deleted"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Product not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Soft-delete a product",
        "tags": [
          "Admin – Products"
        ]
      }
    },
...
    "/reviews/{id}": {
      "delete": {
        "operationId": "ReviewsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Review UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Review deleted"
          },
          "401": {
            "description": "Missing or invalid token"
          },
          "403": {
            "description": "Not the review owner"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Delete your own review",
        "tags": [
          "Reviews"
        ]
      }
    },
    "/admin/reviews": {
      "get": {
        "operationId": "AdminReviewsController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "productId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/ReviewStatus"
            }
          },
          {
            "name": "userId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "schema": {
              "example": "great quality",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedAdminReviewResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "List all reviews with filters (all statuses)",
        "tags": [
          "Admin – Reviews"
        ]
      }
    },
    "/admin/reviews/{id}/status": {
      "patch": {
        "operationId": "AdminReviewsController_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Review UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateReviewStatusDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminReviewResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Approve or reject a review",
        "tags": [
          "Admin – Reviews"
        ]
      }
    },
    "/admin/reviews/{id}": {
      "delete": {
        "operationId": "AdminReviewsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Review UUID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Review deleted"
          },
          "401": {
            "description": "Missing or invalid JWT"
          },
          "403": {
            "description": "Admin role required"
          },
          "404": {
            "description": "Review not found"
          }
        },
        "security": [
          {
            "access_token": []
          }
        ],
        "summary": "Hard-delete a review",
        "tags": [
          "Admin – Reviews"
        ]
      }
    },
...
  },
  "info": {
    "title": "O2Shop API",
    "description": "REST API for the O2Shop e-commerce platform. Public endpoints need no authentication. Customer endpoints require a JWT bearer token (from POST /auth/login). Admin endpoints additionally require the `admin` role.",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3001",
      "description": "Local development"
    }
  ],
  "components": {
    "securitySchemes": {
      "access_token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Paste the access_token value from POST /auth/login."
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "example": [
              "email must be an email"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "error": {
            "type": "string",
            "example": "Bad Request"
          }
        },
        "required": [
          "statusCode",
          "message",
          "error"
        ]
      },
...
      "ProductPhotoResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "url": {
            "type": "string",
            "example": "/uploads/photo.jpg"
          },
          "altText": {
            "type": "string",
            "example": "Front view of product"
          },
          "sortOrder": {
            "type": "number",
            "example": 0
          }
        },
        "required": [
          "id",
          "url",
          "sortOrder"
        ]
      },
...
      "ProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "blue_widget"
          },
          "displayName": {
            "type": "string",
            "example": "Blue Widget"
          },
          "basePrice": {
            "type": "number",
            "example": 29.99
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "description": {
            "type": "object",
            "required": [
              "blocks"
            ],
            "properties": {
              "blocks": {
                "type": "array",
                "items": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "type",
                        "content"
                      ],
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "text"
                          ]
                        },
                        "content": {
                          "type": "string",
                          "example": "This is a paragraph."
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "type",
                        "items"
                      ],
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "points"
                          ]
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "Feature one",
                            "Feature two"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "rating": {
            "type": "object",
            "nullable": true,
            "example": 4.5
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductPhotoResponseDto"
            }
          },
          "defaultVariant": {
            "$ref": "#/components/schemas/ProductVariantResponseDto"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantResponseDto"
            }
          },
          "collection": {
            "$ref": "#/components/schemas/CollectionSummaryDto"
          },
          "category": {
            "$ref": "#/components/schemas/CategorySummaryDto"
          },
          "subCategory": {
            "$ref": "#/components/schemas/SubCategorySummaryDto"
          }
        },
        "required": [
          "id",
          "name",
          "displayName",
          "basePrice",
          "currency",
          "photos",
          "variants",
          "category",
          "subCategory"
        ]
      },
      "PaginatedProductResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductResponseDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "data"
        ]
      },
      "AdminProductResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "example": "blue_widget"
          },
          "displayName": {
            "type": "string",
            "example": "Blue Widget"
          },
          "basePrice": {
            "type": "number",
            "example": 29.99
          },
          "currency": {
            "type": "string",
            "example": "USD"
          },
          "description": {
            "type": "object",
            "required": [
              "blocks"
            ],
            "properties": {
              "blocks": {
                "type": "array",
                "items": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "type",
                        "content"
                      ],
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "text"
                          ]
                        },
                        "content": {
                          "type": "string",
                          "example": "This is a paragraph."
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "type",
                        "items"
                      ],
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "points"
                          ]
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "Feature one",
                            "Feature two"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "rating": {
            "type": "object",
            "nullable": true,
            "example": 4.5
          },
          "photos": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductPhotoResponseDto"
            }
          },
          "defaultVariant": {
            "$ref": "#/components/schemas/ProductVariantResponseDto"
          },
          "variants": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProductVariantResponseDto"
            }
          },
          "collection": {
            "$ref": "#/components/schemas/CollectionSummaryDto"
          },
          "category": {
            "$ref": "#/components/schemas/CategorySummaryDto"
          },
          "subCategory": {
            "$ref": "#/components/schemas/SubCategorySummaryDto"
          },
          "isPublished": {
            "type": "boolean",
            "example": true
          },
          "deletedAt": {
            "type": "object",
            "nullable": true,
            "example": null
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "displayName",
          "basePrice",
          "currency",
          "photos",
          "variants",
          "category",
          "subCategory",
          "isPublished",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "UpdateProductDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "blue_widget",
            "pattern": "^[a-z0-9_]+$",
            "description": "Unique URL-safe identifier (lowercase, underscores)"
          },
          "displayName": {
            "type": "string",
            "example": "Blue Widget"
          },
          "collectionId": {
            "type": "string",
            "format": "uuid",
            "description": "ID of the collection to assign"
          },
          "categoryId": {
            "type": "string",
            "format": "uuid"
          },
          "subCategoryId": {
            "type": "string",
            "format": "uuid"
          },
          "basePrice": {
            "type": "number",
            "minimum": 0,
            "example": 29.99
          },
          "currency": {
            "type": "string",
            "minLength": 3,
            "maxLength": 3,
            "example": "USD"
          },
          "description": {
            "type": "object",
            "properties": {
              "blocks": {
                "type": "array",
                "items": {
                  "oneOf": [
                    {
                      "type": "object",
                      "required": [
                        "type",
                        "content"
                      ],
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "text"
                          ]
                        },
                        "content": {
                          "type": "string",
                          "example": "This is a paragraph."
                        }
                      }
                    },
                    {
                      "type": "object",
                      "required": [
                        "type",
                        "items"
                      ],
                      "properties": {
                        "type": {
                          "type": "string",
                          "enum": [
                            "points"
                          ]
                        },
                        "items": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          },
                          "example": [
                            "Feature one",
                            "Feature two"
                          ]
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "isPublished": {
            "type": "boolean",
            "default": false,
            "example": false
          }
        }
      },
...
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected"
        ]
      },
      "AdminReviewResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "productId": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "displayName": {
            "type": "string",
            "example": "Jane D."
          },
          "rating": {
            "type": "number",
            "minimum": 1,
            "maximum": 10,
            "example": 8
          },
          "content": {
            "type": "string",
            "example": "Great quality, fast shipping!"
          },
          "photoUrls": {
            "example": [
              "https://example.com/photo.jpg"
            ],
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          },
          "email": {
            "type": "string",
            "example": "user@example.com"
          },
          "status": {
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewStatus"
              }
            ]
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "productId",
          "displayName",
          "rating",
          "content",
          "createdAt",
          "email",
          "status",
          "updatedAt"
        ]
      },
      "PaginatedAdminReviewResponseDto": {
        "type": "object",
        "properties": {
          "total": {
            "type": "number",
            "example": 100
          },
          "page": {
            "type": "number",
            "example": 1
          },
          "limit": {
            "type": "number",
            "example": 20
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AdminReviewResponseDto"
            }
          }
        },
        "required": [
          "total",
          "page",
          "limit",
          "data"
        ]
      },
...
      "UpdateReviewStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "example": "approved",
            "allOf": [
              {
                "$ref": "#/components/schemas/ReviewStatus"
              }
            ]
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}